+++ spring-oxygen-flysql/src/main/java/com/isxcode/oxygen/flysql/parse/SqlServerSqlValue.java
@@ -1,0 +1,5 @@
+package com.isxcode.oxygen.flysql.parse;
+
+public class SqlServerSqlValue implements SqlValue {
+
+}